+++ src/main/java/com/stamp/api/auth/employer/dto/request/LoginEmployerReq.java
@@ -1,3 +1,3 @@
-package com.stamp.api.auth.dto.request;
+package com.stamp.api.auth.employer.dto.request;
 
 public record LoginEmployerReq(String email, String password) {}
