+++ src/main/java/com/stamp/api/auth/employer/infra/oauth/OAuthMemberClient.java
@@ -1,6 +1,6 @@
-package com.stamp.api.auth.infra.oauth;
+package com.stamp.api.auth.employer.infra.oauth;
 
-import com.stamp.api.auth.dto.request.SocialLoginEmployerReq;
+import com.stamp.api.auth.employer.dto.request.SocialLoginEmployerReq;
 import com.stamp.api.employeruser.entity.EmployerUser;
 
 public interface OAuthMemberClient {
