+++ src/main/java/com/stamp/api/auth/employer/infra/oauth/OAuthRequestProvider.java
@@ -1,4 +1,4 @@
-package com.stamp.api.auth.infra.oauth;
+package com.stamp.api.auth.employer.infra.oauth;
 
 public interface OAuthRequestProvider {
   ProviderType supportType();
