+++ src/main/java/com/stamp/api/auth/employer/infra/oauth/OAuthUserDetails.java
@@ -1,0 +1,5 @@
+package com.stamp.api.auth.employer.infra.oauth;
+
+import com.stamp.api.auth.employer.entity.OAuthId;
+
+public record OAuthUserDetails(OAuthId oAuthId, String email) {}
