--- src/main/java/com/stamp/api/auth/infra/oauth/OAuthUserDetails.java
@@ -1,5 +1,0 @@
-package com.stamp.api.auth.infra.oauth;
-
-import com.stamp.api.auth.entity.OAuthId;
-
-public record OAuthUserDetails(OAuthId oAuthId, String email) {}
