--- src/main/java/com/stamp/api/auth/service/AuthService.java
@@ -1,8 +1,0 @@
-package com.stamp.api.auth.service;
-
-import com.stamp.api.auth.dto.request.LoginEmployerReq;
-import com.stamp.api.auth.dto.response.LoginRes;
-
-public interface AuthService {
-  LoginRes login(LoginEmployerReq loginEmployerReq);
-}
