+++ src/main/java/com/stamp/api/employeruser/entity/EmployerUser.java
@@ -1,6 +1,6 @@
 package com.stamp.api.employeruser.entity;
 
-import com.stamp.api.auth.entity.OAuthId;
+import com.stamp.api.auth.employer.entity.OAuthId;
 import jakarta.annotation.Nullable;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
