+++ src/main/java/com/stamp/api/employeruser/repository/EmployerUserRepository.java
@@ -1,6 +1,6 @@
 package com.stamp.api.employeruser.repository;
 
-import com.stamp.api.auth.entity.OAuthId;
+import com.stamp.api.auth.employer.entity.OAuthId;
 import com.stamp.api.employeruser.entity.EmployerUser;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
