+++ src/main/java/com/stamp/global/config/HttpInterfaceConfig.java
@@ -1,6 +1,6 @@
 package com.stamp.global.config;
 
-import com.stamp.api.auth.infra.oauth.kakao.KakaoApiClient;
+import com.stamp.api.auth.employer.infra.oauth.kakao.KakaoApiClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
